adding seeders to test api in progress

+++ database/migrations/2025_01_19_125805_create_board_role_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('board_role', function (Blueprint $table) {
             $table->id();
-            $table->integer('role_id');
-            $table->integer('board_id');
+            $table->unsignedBigInteger('role_id');
+            $table->unsignedBigInteger('board_id');
             $table->timestamps();
         });
     }
